qcacmn: Do not use preCAC timer check for Rolling CAC
While checking for the radar source, the status of preCAC timer
is used to check if rolling CAC is active or not.

However, rolling CAC uses a different timer which results in
radar during rolling CAC to be treated as radar in home channel.
Check for preCAC timer only if agile preCAC is enabled.

CRs-Fixed: 2811313
Change-Id: I1f41e4fb83213abb8fd93531174063fad339f3b7
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>