qcacmn: Add zeroCAC(legacy) check before assigning cfreq2
While getting the bonded channels during radar, if preCAC timer
is running, we use the zeroCAC frequency as the secondary frequency.

Consider a device with agile DFS enabled and the home channel as
160MHz/80P80MHz. If radar is injected on the secondary segment of
the home channel, since agile preCAC is running, the zeroCAC frequency
is considered as the secondary center. This results in wrong
channels being added to NOL/failure (depending on the zeroCAC channels).

Use the zeroCAC frequency as secondary segment center during radar,
only if legacy preCAC is enabled.

Change-Id: I9a99b1c9968e622ffe55c662fd21586cfc587281
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>