qcacld-3.0: Add helper function to typecast channel band parameter
Prior to the kernel version4.7 struct ieee80211_channel used enum
ieee80211_band. Beginning with kernel version4.7 struct ieee80211_channel
use enum nl80211_band. So there is an issue of typecasting of the channel
band parameter.

To resolve the above issue a helper function is added to typecast the band
parameter according to the kernel version.

Change-Id: I1af692bb7cae5ab2e71556128ce672c6b145f92f
CRs-Fixed: 2815201
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>