Revert "disp: msm: sde: fix race between disable commit and vblank work"
This is a partial revert of commit:
	045975ff495be49b3f9c74f86b6d1ce7b8a21bf4.

The enc->crtc link cannot be relied upon outside of the display
thread context since the DRM legacy modeset API will detach and
re-attach this connection on every commit.

The logic to flush the event thread in all cases during the CRTC
disable is kept from the original patch.

Change-Id: I2cd4ae47589bb397c72de0e9deba9b0c710d379a
Signed-off-by: Steve Cohen <cohens@codeaurora.org>