qcacld-3.0: Add a check for invalid channel number
If the channel bandwidth is not 80+80, then the channel freq segment1
of the current session will become 0. When api is called to get freq from
channel num it shows error msg invalid channel num 0.

To avoid the above issue, add a check for invalid channel number before
calling the api to get freq.

Change-Id: I7bd0748be04181d4e59935b0f9111d582ca31e3c
CRs-Fixed: 2820653
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>