qcacld-3.0: Don't disable btm offload for open mode connection
Currently driver disables btm offload to firmware if connected AP
supports MBO-OCE and the connection is not PMF enabled. But the
pmf will always be disabled for open mode connection, so BTM
offload is disabled with open mode MBO-OCE enabled AP always.

Check if the connection is open mode connection and allow
BTM offload config.

Change-Id: I8fe9b43c07700dca02cf7bb874e1c5c509b120f7
CRs-Fixed: 2831463
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>